Rename, clean up, add .csproj for CI verification. Add using declaration, etc.
