+++ samples/snippets/csharp/VS_Snippets_CLR/conceptual.disposable/cs/TryFinallyGenerated.cs
@@ -1,15 +1,14 @@
-﻿// <Snippet3>
-using System;
-using System.IO;
+﻿using System.IO;
 
-class Example
+class TryFinallyGenerated
 {
     static void Main()
     {
-        char[] buffer = new char[50];
-        var streamReader = new StreamReader("file1.txt");
+        var buffer = new char[50];
+        StreamReader? streamReader = null;
         try
         {
+            streamReader = new StreamReader("file1.txt");
             int charsRead = 0;
             while (streamReader.Peek() != -1)
             {
@@ -21,11 +20,8 @@ class Example
         }
         finally
         {
-            if (streamReader != null)
-            {
-                ((IDisposable)streamReader).Dispose();
-            }
+            // If non-null, call the object's Dispose method.
+            streamReader?.Dispose();
         }
     }
 }
-// </Snippet3>
