--- samples/snippets/visualbasic/VS_Snippets_CLR/conceptual.disposable/vb/using1.vb
@@ -1,21 +1,0 @@
-﻿' Visual Basic .NET Document
-Option Strict On
-
-' <Snippet1>
-Imports System.IO
-
-Module Example
-    Public Sub Main()
-        Dim buffer(49) As Char
-        Using s As New StreamReader("File1.txt")
-            Dim charsRead As Integer
-            Do While s.Peek() <> -1
-                charsRead = s.Read(buffer, 0, buffer.Length)
-                ' 
-                ' Process characters read.
-                '
-            Loop
-        End Using
-    End Sub
-End Module
-' </Snippet1>
